Push config file locating to modules
Each module now accepts the name of a config file and will
internally compute the path to that file based on the pipelines it
executes. For example, to run on a SLURM cluster, a tester would
specify 'slurm.config'. Each module would the look in the proper
directory for the pipeline it runs. If the module runs EST, it
would define the config path as 'conf/est/docker.config'. A
generate SSN pipeline would similarly identify
'conf/generatessn/docker.config' as the file to use.