Create annotated mapping table
The annotated mapping tables are now created in the color SSN pipeline.  The two
tables that are current created are the mapping table which contains metadata
for every UniProt ID (expanded from UniRef/RepNode IDs), and a second table
containing a list of IDs that are SwissProt sequences with the description.
A color mapping file is now created by color_xgmml.pl. Also, since the
pipeline requires access to a FASTA database to retrieve the sequences,
the test was updated to reflect that.