Create numbering and mapping files for clusters and singletons
Several steps are needed to obtain the information for the pipeline (and future
derivatives).  During the ID list retrieval process, cluster sizes are computed.
During the cluster computation process, singletons are saved as well as information
on the relationship between clusters numbered by number of sequences vs number
of nodes.